Copy step/degree conversions from pupil_imsoft
To not to have dependency to pupil_imsoft for such a small thing,
copy relevant angle conversion parts in rotary_encoders.py which
now also contains DEFAULT_STEPS_PER_ROTATION variable.

Imsoft should be set to enquire from user and record this parameter
in all data because it can be that the rotaion encoders are changed
over time or there are multiple setups with different encoder
steps per full rotation.