Fix accidental pupildir pollution (experimental)
If user accidentally selects a wrong directory in the GUI,
pupildir got polluted by any folders that were on that dir
because MAnalyser subfolders had to be created.

Now it is changed so that if no_data_load is specified for the
MAnalyser, do not create these folders.

To be seen if this causes more trouble that it tries to fix.