Rework field_error
Fixes temporal solution made in the commit
32ecb46e77c3bf57340fbd4e898de512ac68a0ff; Now it doesn't matter
anymore if the points overlap perfectly; If so, only perfectly
overlapping points are taken.

Moreover, instead of taking the distance weighted mean of 10 closest
vectors and calculating error for that, take 10 closest vectors
and calculate 10 errors and take weighted mean of the errors based
on the distance from the point where the error is calculated.
This gives more reliable results on the left/right midline
where vectors from left and right eyes overlap.