Fix get_imaging_frequency to allow image_folder=None
Some plotting functions want to know the imaging frequency even
if they operate on the specimen level - meaning that for them
image_folder=None.

This commit changes the get_imaging_frequency to use the first
imaging folder that gives out a fs and report it.

This of course fails if for one specimen there are image folders
with multiple imaging frequencies....