Fix tkgui ROI/image plotting with multiple resolutions
Many different resolutions for the same specimen to cause
bugged view of the ROI/images because only matplotlib's set_data
was used, and the figure was not fully replotted.

This commit changes to use CanvasPlotter's imshow method instead
where the issue is already taken care of.

Axis ticks are lost until CanvasPlotter gains a feature to have
them remained.